+++ src/test/org/firebirdsql/jdbc/TestFBResultSetMetaData.java
@@ -20,6 +20,10 @@ package org.firebirdsql.jdbc;
 
 import org.firebirdsql.common.DdlHelper;
 import org.firebirdsql.common.rules.UsesDatabase;
+import org.firebirdsql.encodings.EncodingFactory;
+import org.firebirdsql.gds.ng.DefaultDatatypeCoder;
+import org.firebirdsql.gds.ng.fields.RowDescriptor;
+import org.firebirdsql.gds.ng.fields.RowDescriptorBuilder;
 import org.firebirdsql.util.FirebirdSupportInfo;
 import org.junit.BeforeClass;
 import org.junit.ClassRule;
@@ -28,10 +32,14 @@ import org.junit.Test;
 import javax.sql.rowset.CachedRowSet;
 import javax.sql.rowset.RowSetFactory;
 import javax.sql.rowset.RowSetProvider;
+import java.nio.charset.StandardCharsets;
 import java.sql.*;
+import java.util.Collections;
 import java.util.Properties;
 
 import static org.firebirdsql.common.FBTestProperties.*;
+import static org.firebirdsql.gds.ISCConstants.SQL_DOUBLE;
+import static org.firebirdsql.gds.ISCConstants.SQL_FLOAT;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.fail;
 import static org.junit.Assume.assumeTrue;
@@ -389,4 +397,19 @@ public class TestFBResultSetMetaData {
             }
         }
     }
+
+    @Test
+    public void getPrecision_connectionLessResultSet_shouldSucceedWithoutException_730() throws Exception {
+        RowDescriptor rowDescriptor = new RowDescriptorBuilder(2,
+                DefaultDatatypeCoder.forEncodingFactory(EncodingFactory.createInstance(StandardCharsets.UTF_8)))
+                .at(0).simple(SQL_FLOAT, 4, "TEST", "FLOAT").addField()
+                .at(1).simple(SQL_DOUBLE, 8, "TEST", "DOUBLE").addField()
+                .toRowDescriptor();
+        ResultSet rs = new FBResultSet(rowDescriptor, Collections.emptyList());
+        ResultSetMetaData rsmd = rs.getMetaData();
+
+        assertEquals(24, rsmd.getPrecision(1));
+        assertEquals(53, rsmd.getPrecision(2));
+    }
+
 }